# Core/protocol/omni2fa.openapi.yaml
openapi: 3.1.0
info:
  title: Omni2FA API
  version: 0.6.0
  summary: Cross-stack HTTP contract for multi-method 2FA.
  description: |
    The HTTP contract that any backend implementing Omni2FA must satisfy, and
    that any frontend implementing Omni2FA expects. The same contract is the
    integration point for cross-stack combinations (React + .NET, React + Python,
    Angular + .NET, …).

    **This document is the source of truth.** Backend implementations (.NET,
    future Python) and frontend implementations (React, future Vue/Angular)
    are generated from or verified against this file.

    ## Scope per milestone

    | Milestone | Endpoints included in this spec |
    |-----------|---------------------------------|
    | v0.0 (current, pre-alpha) | Scaffolding — TOTP enrollment + login challenge, methods list/delete. |
    | v0.1 | Same as v0.0 but with implementations on .NET + React. |
    | v0.2 | + Email OTP endpoints (`/enroll/email/*`). |
    | v0.3 | + WebAuthn endpoints (`/enroll/webauthn/*`). |
    | v0.4 | + Recovery codes (`/recovery-codes/*`, returned in enrollment responses). |
    | v1.0 | Frozen — breaking changes from here on bump major. |

    See `docs/FLOWS.md` for sequence diagrams and `Core/protocol/ERROR_CODES.md`
    for the full error code catalogue.

    ## Versioning

    `info.version` matches the Omni2FA library milestone version. Packages on
    the same `MAJOR.MINOR.*` are compatible — see `docs/ROADMAP.md` "Versioning
    model".

...
servers:
...
paths:
...
  /challenge/start:
    post:
      tags: [challenge]
      operationId: startChallenge
      summary: Start verification of a chosen method.
      description: |
        After the host verifies the user's password and issues a pre-auth token,
        the frontend calls this with the picked method. Backend prepares
        method-specific challenge state — see response shape per type.

        For **TOTP** — this is a no-op acknowledgment; the user enters the code
        from their authenticator and the frontend goes straight to `/verify`.

        For **Email** (v0.2+) — backend sends an OTP email and returns confirmation.

        For **WebAuthn** (v0.3+) — backend returns assertion request options
        for `navigator.credentials.get()`.
      security:
        - PreAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeStartRequest'
      responses:
        '200':
          description: Challenge prepared. Response shape depends on method type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/PreAuthInvalidOrExpired'
        '404':
          description: Method not found, or does not belong to current pre-auth user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /challenge/resend:
    post:
      tags: [challenge]
      operationId: resendChallenge
      summary: Re-send the login OTP for a method that delivers codes out-of-band (Email).
      description: |
        For **Email** — re-sends the one-time code for the active login challenge of the
        given method, subject to a resend cooldown. For methods that don't deliver codes
        (TOTP, WebAuthn) this is not applicable and returns `400 VALIDATION_FAILED`.
      security:
        - PreAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeResendRequest'
      responses:
        '200':
          description: A fresh code was sent (or the previous one re-sent).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/PreAuthInvalidOrExpired'
        '404':
          description: No active challenge for the given method under the current pre-auth user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /challenge/recovery-code:
    post:
      tags: [challenge]
      operationId: verifyRecoveryCode
      summary: Complete login with a one-time recovery code instead of a 2FA method.
      description: |
        Method-agnostic fallback. Consumes the recovery code (one-time use) and, on success,
        returns the verified `userId` exactly like `/challenge/verify`.
      security:
        - PreAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryCodeVerifyRequest'
      responses:
        '200':
          description: Recovery code accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifySuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Code invalid, already used, or pre-auth invalid/expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /challenge/verify:
    post:
      tags: [challenge]
      operationId: verifyChallenge
      summary: Verify the user's response to a challenge.
      description: |
        Final step of the 2FA login flow. On success, the host application
        receives the `userId` and is responsible for minting its session JWT.

        Omni2FA does **not** issue session tokens — only verification results.
      security:
        - PreAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeVerifyRequest'
      responses:
        '200':
          description: Verification succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifySuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Code invalid, pre-auth invalid/expired, or challenge consumed/missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ------------------------------------------------------------------
  #  enrollment — Email OTP
  # ------------------------------------------------------------------
  /enroll/email/start:
    post:
      tags: [enroll-email]
      operationId: startEmailEnrollment
      summary: Begin Email OTP enrollment.
      description: |
        The host supplies the destination email address — Omni2FA does not read it from a
        claim and does not own address verification (that is the host's responsibility). The
        server issues a one-time code, persists it as a pending enrollment, and emails it.
      security:
        - HostSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailEnrollStartRequest'
      responses:
        '200':
          description: Enrollment started, OTP emailed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailEnrollStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: An Email method is already enrolled for this user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /enroll/email/confirm:
    post:
      tags: [enroll-email]
      operationId: confirmEmailEnrollment
      summary: Confirm Email OTP enrollment by verifying the emailed code.
      security:
        - HostSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailEnrollConfirmRequest'
      responses:
        '200':
          description: Email method enrolled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MethodCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No pending Email enrollment for this user, or it has expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /enroll/email/resend:
    post:
      tags: [enroll-email]
      operationId: resendEmailEnrollment
      summary: Re-send the enrollment OTP, subject to a resend cooldown.
      security:
        - HostSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailEnrollResendRequest'
      responses:
        '200':
          description: A fresh code was sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailEnrollStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No pending Email enrollment for this user, or it has expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ------------------------------------------------------------------
  #  enrollment — WebAuthn
  # ------------------------------------------------------------------
...
  /enroll/totp/start:
    post:
      tags: [enroll-totp]
      operationId: startTotpEnrollment
      summary: Begin TOTP enrollment.
      description: |
        Generates a fresh secret, persists it as a pending enrollment,
        and returns the otpauth:// URI plus the base32 secret. The frontend
        renders the URI as a QR and shows the secret for manual entry.
      security:
        - HostSession: []
      responses:
        '200':
          description: Enrollment ceremony started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollStartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: TOTP method already enrolled for this user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /enroll/totp/confirm:
    post:
      tags: [enroll-totp]
      operationId: confirmTotpEnrollment
      summary: Confirm TOTP enrollment by verifying a code.
      description: |
        User scans the QR and submits the first 6-digit code from their
        authenticator. Backend validates the code against the pending secret
        and persists the method on success.
      security:
        - HostSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TotpEnrollConfirmRequest'
      responses:
        '200':
          description: TOTP enrolled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MethodCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No pending TOTP enrollment for this user, or it has expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ------------------------------------------------------------------
  #  recovery codes
  # ------------------------------------------------------------------
...
components:
...
  securitySchemes:
    HostSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The host application's full session token. Omni2FA does not issue
        this — your existing auth does. Endpoints under `methods` and
        `enroll-*` require this. The user has fully logged in.
    PreAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived (default 3-5 min) pre-auth token issued by Omni2FA after
        the host has verified the user's password but before 2FA verification
        completes. Used exclusively by `/challenge/*` endpoints.
        Token is invalidated immediately after a successful `/challenge/verify`.

  # ------------------------------------------------------------------
  #  Parameters
  # ------------------------------------------------------------------
...
  schemas:
...
    TwoFactorMethodType:
      type: string
      enum: [Totp, Email, WebAuthn]
      description: |
        The type of 2FA factor. Currently only `Totp` is implemented (v0.1).
        `Email` lands in v0.2, `WebAuthn` in v0.3.

    # --- Method DTOs ---
...
    ChallengeStartRequest:
      type: object
      required: [methodId]
      properties:
        methodId:
          type: string
          format: uuid
          description: ID of the method the user picked from `availableMethods`.
    ChallengeStartResponse:
      type: object
      description: |
        Method-specific challenge payload. `type` discriminates the shape.
        For TOTP — empty payload (just confirms challenge started); `expiresAt` and
        `resendAvailableAt` are null. For Email — an OTP was sent; `expiresAt` is when
        it stops validating and `resendAvailableAt` is when a resend becomes allowed.
        WebAuthn assertion options will be added in v0.3.
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/TwoFactorMethodType'
        expiresAt:
          type: [string, "null"]
          format: date-time
          description: For Email — when the sent code stops validating. Null for TOTP.
        resendAvailableAt:
          type: [string, "null"]
          format: date-time
          description: For Email — earliest time a resend is permitted. Null for TOTP.
        optionsJson:
          type: [string, "null"]
          description: |
            For WebAuthn — `PublicKeyCredentialRequestOptions` JSON to pass to
            `navigator.credentials.get()`. Null for TOTP and Email.
    ChallengeResendRequest:
      type: object
      required: [methodId]
      properties:
        methodId:
          type: string
          format: uuid
          description: The method whose active login OTP should be re-sent.
    ChallengeVerifyRequest:
      type: object
      required: [methodId]
      properties:
        methodId:
          type: string
          format: uuid
        code:
          type: [string, "null"]
          description: For TOTP and Email — the 6-digit numeric code. Null for WebAuthn.
        assertionResponseJson:
          type: [string, "null"]
          description: For WebAuthn — the JSON from `navigator.credentials.get()`. Null for TOTP and Email.
    VerifySuccessResponse:
      type: object
      required: [verified, userId]
      properties:
        verified:
          type: boolean
          description: Always `true` on this response. Failures use `4xx` with `ErrorResponse`.
        userId:
          type: string
          maxLength: 64
          description: |
            Verified user's identifier, stringified. The host application casts
            this to its native userId type (Guid/long/int/email) and uses it to
            mint its final session JWT/cookie. Omni2FA does not issue session tokens.

    # --- Email enrollment ---
    EmailEnrollStartRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: |
            Destination address the OTP is sent to, supplied by the host. Omni2FA does not
            derive it from a claim and does not verify address ownership — that is host policy.
          example: 'alice@example.com'
    EmailEnrollStartResponse:
      type: object
      required: [enrollmentId, expiresAt, resendAvailableAt]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Identifier for the pending enrollment ceremony. Pass back in `confirm` / `resend`.
        expiresAt:
          type: string
          format: date-time
          description: When the emailed code stops validating.
        resendAvailableAt:
          type: string
          format: date-time
          description: Earliest time a resend is permitted (cooldown).
    EmailEnrollConfirmRequest:
      type: object
      required: [enrollmentId, code]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Value returned from `/enroll/email/start`.
        code:
          type: string
          description: The numeric code from the enrollment email.
          example: '123456'
        name:
          type: [string, "null"]
          description: Optional human-readable label for this method.
    EmailEnrollResendRequest:
      type: object
      required: [enrollmentId]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Value returned from `/enroll/email/start`.

    # --- WebAuthn enrollment ---
...
    TotpEnrollStartResponse:
      type: object
      required: [enrollmentId, otpAuthUri, secret]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Identifier for the pending enrollment ceremony. Pass back in `confirm`.
        otpAuthUri:
          type: string
          description: |
            `otpauth://` URI for authenticator app QR code. Format:
            `otpauth://totp/{Issuer}:{accountLabel}?secret={base32}&issuer={Issuer}&algorithm=SHA1&digits=6&period=30`.
          example: 'otpauth://totp/MyApp:alice@example.com?secret=JBSWY3DPEHPK3PXP&issuer=MyApp&algorithm=SHA1&digits=6&period=30'
        secret:
          type: string
          description: Base32-encoded TOTP secret. Shown to the user as a fallback to QR scanning.
          example: 'JBSWY3DPEHPK3PXP'
    TotpEnrollConfirmRequest:
      type: object
      required: [enrollmentId, code]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Value returned from `/enroll/totp/start`.
        code:
          type: string
          description: 6-digit code from the user's authenticator app.
          example: '123456'
        name:
          type: [string, "null"]
          description: Optional human-readable label for this method.
    MethodCreatedResponse:
      type: object
      required: [methodId]
      properties:
        methodId:
          type: string
          format: uuid
          description: Identifier of the newly created method.
        recoveryCodes:
          type: [array, "null"]
          items:
            type: string
          description: |
            Present only when this enrollment generated recovery codes (i.e. it was the user's
            first method). Plaintext, shown once — the frontend must display and have the user
            save them. Null on every subsequent enrollment.
...
    RecoveryCodeVerifyRequest:
      type: object
      required: [recoveryCode]
      properties:
        recoveryCode:
          type: string
          description: A single unused recovery code. Dashes/spaces and case are normalized server-side.
          example: 'A1B2-C3D4-E5'

    # --- Errors ---
    ErrorResponse:
      type: object
      required: [code, message]
      description: |
        Uniform error envelope returned by every non-2xx response. Frontends
        switch on `code` (machine-readable, stable). `message` is informational
        and may be localized.
        See `Core/protocol/ERROR_CODES.md` for the full catalogue.
      properties:
        code:
          type: string
          description: Machine-readable error code, stable across versions.
          example: 'INVALID_CODE'
        message:
          type: string
          description: Human-readable message. May be localized by host.
          example: 'The code you entered is invalid.'
        details:
          type: [object, "null"]
          additionalProperties: true
          description: Optional structured context. Shape depends on `code`.

  # ------------------------------------------------------------------
  #  Reusable responses
  # ------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreAuthInvalidOrExpired:
      description: Pre-auth token is missing, malformed, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: |
        Rate limit exceeded. Default policy is 20 attempts per minute per IP
        on verify endpoints — see `docs/CODE_STYLE.md` rule 5 + ROADMAP v0.6.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        Retry-After:
          description: Seconds until the next attempt is allowed.
          schema:
            type: integer